% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{calc_proliferation_scores}
\alias{calc_proliferation_scores}
\title{Calculate proliferation score}
\usage{
calc_proliferation_scores(centered_input)
}
\arguments{
\item{centered_input}{numeric \link{matrix}; 50 rows (PAM50 genes) x N columns (sample). See \link{preprocess_input}.}
}
\value{
named \link{numeric} vector of per sample prolif. score
}
\description{
\code{calc_proliferation_scores} calculates a single numeric score per sample based on a subset of 11 PAM50 genes.
}
