% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{calc_ror_scores}
\alias{calc_ror_scores}
\title{Calculate Rick of Relapse (ROR) scores}
\usage{
calc_ror_scores(pam50_assignment, prolif_score, tumor_size = NA)
}
\arguments{
\item{pam50_assignment}{named \link{numeric}; single row from \link{assign_subtypes}}

\item{prolif_score}{\link{numeric}; scaler value. See \link{calc_proliferation_scores}}

\item{tumor_size}{OPTIONAL \link{logical}; indicator of tumor size > 2cm. If \link{NA}, values depending on tumor_size will be \link{NA_real_}.}
}
\value{
1 row \link{data.frame} with alternating ROR scores, group, per model type
}
\description{
\code{calc_ror_scores} returns a named list of ROR scores for 1 sample
}
\details{
Input is the set of PAM50 centroid correlation
}
