% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{es}
\alias{es}
\title{TCGA BRCA RNA-seq data (partial).}
\format{An object of class \code{ExpressionSet} with 20133 rows and 50 columns.}
\usage{
es
}
\description{
TCGA BRCA RNA-seq data (partial).
}
\keyword{datasets}
