% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{extract_PAM50}
\alias{extract_PAM50}
\title{Extract PAM50 genes}
\usage{
extract_PAM50(df, an = NULL)
}
\arguments{
\item{df}{expression dataframe, rownames has to be EntrezID (Gene ID)}

\item{an}{annotation dataframe. Optional, if passed in, those 50 gene info will be extracted.}
}
\value{
es.50 dataframe with extracted PAM50 gene symbols as rownames

an.50 (optional) if an is passed in, it will return an.50 dataframe that contains extracted PAM50 gene info
}
\description{
Extract PAM50 genes
}
