% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{filter_sort}
\alias{filter_sort}
\title{Filter and sort input expression matrix to match PAM50 centroids}
\usage{
filter_sort(input)
}
\arguments{
\item{input}{\link{matrix}; numeric matrix of gene expression with PAM50 gene symbols as row names.}
}
\value{
50 x N_sample numeric matrix with row order matching \link{pam50_centroids}.
}
\description{
Filter and sort input expression matrix to match PAM50 centroids
}
