% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50.R
\name{getPAM50}
\alias{getPAM50}
\title{PAM50 prediction via PAMR}
\usage{
getPAM50(df)
}
\arguments{
\item{df}{dataframe, rows are genes columns are samples. Row names has to be EntrezID.}
}
\description{
After getting the rank values, data will be passed into `getPAM50` which uses
a trained PAMR model for prediction. This will give a single PAM50 subtype prediction
for each sample, it also provides the prediction probability for all 5 subtypes
per patient.
}
