% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50.R
\name{getPAM50rank}
\alias{getPAM50rank}
\title{Convert expression values to Rank}
\usage{
getPAM50rank(df)
}
\arguments{
\item{df}{dataframe, rows are genes columns are samples. Row names has to be EntrezID.}
}
\description{
`getPAM50rank(df)` first extracts PAM50 genes based on the EntrezID. Therefore,
the input dataframe `df` (usually the assayData from expressionSet;
rows are genes, and columns are samples) has to have rownames assigned to
EntrezID (Gene ID). After extracting 50 genes, it will then rank those 50 genes
based on the values per sample, and then finally normalized the rank values
by the number of genes in the dataframe (usually 50 genes) .
}
\author{
Andrew Chang
}
