% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50.R
\name{PAM50}
\alias{PAM50}
\title{PAM50 main function}
\usage{
PAM50(df, cutoff = 0)
}
\arguments{
\item{df}{dataframe, rows are genes columns are samples. Row names has to be EntrezID.}

\item{cutoff}{numeric (0-1, default = 0), threshold value for calling the prediction.
0 means no cutoff, predicted subtype is whichever subtype gives the highest probability.
0.8 means the predicted subtype has to have probability > 0.8, otherwise 'NA' is the label.}
}
\description{
This is what you need for getting PAM50 prediction for each sample. It takes
RNA-seq values (raw or CPM) and returns PAM50 subtype for each sample.
}
\examples{
library(Biobase)
library(tidyverse)
library(pamr)

df <- data.frame(assayData(es)$exprs)
df <- df[,1:10] ## select 10 samples
annotation <- fData(es)

rownames(df) <- annotation$EntrezID

df_pred <- PAM50(df)
knitr::kable(df_pred)
}
