% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50_annot}
\alias{pam50_annot}
\title{PAM50 annotations}
\format{
A \link{data.frame} with 50 rows (genes) and 4 columns (annotations).
}
\usage{
pam50_annot
}
\description{
A dataset of PAM50 gene names with IDs and proliferation status
}
\keyword{datasets}
