% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50_centroids}
\alias{pam50_centroids}
\title{PAM50 centroids read from a stone tablet}
\format{
A numeric \link{matrix} with 50 rows (genes) and 5 columns (molecular subtypes)
}
\usage{
pam50_centroids
}
\description{
A dataset containing PAM50 centroids from the original Parker et al. (2009) publication. Availble at: \url{https://web.archive.org/web/20130304090315/https://genome.unc.edu/pubsup/breastGEO/pam50_centroids.txt}
}
\keyword{datasets}
