% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50_palette}
\alias{pam50_palette}
\title{PAM50 palette}
\format{
A named \link{list} of PAM50 subtype hex codes.
}
\usage{
pam50_palette
}
\description{
The 'canonical' PAM50 subtype colors
}
\keyword{datasets}
