% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50_testdata}
\alias{pam50_testdata}
\title{PAM50 test data}
\format{
A numeric \link{matrix} with 50 rows (genes) and 200 columns (samples).
}
\usage{
pam50_testdata
}
\description{
A dataset containing PAM50 gene expression profiles from a microarray panel for 200 subjects. Possibly synthetic data. Available from \url{https://web.archive.org/web/20130302114722/https://genome.unc.edu/pubsup/breastGEO/PAM50.zip}
}
\keyword{datasets}
