% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50_testdata_assignments}
\alias{pam50_testdata_assignments}
\title{PAM50 test data assignments}
\format{
A \link{data.frame} with 200 rows (samples) and 6 columns (cors + assign).
}
\usage{
pam50_testdata_assignments
}
\description{
A dataset containing PAM50 subtype correlations and assignments for the 200 samples from \code{pam50_testdata} as originally called by Parker et al. Availble from \url{https://web.archive.org/web/20130302114722/https://genome.unc.edu/pubsup/breastGEO/PAM50.zip}.
}
\keyword{datasets}
