% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pam50_testdata_samplesheet}
\alias{pam50_testdata_samplesheet}
\title{Test data samplesheet}
\format{
A \link{data.frame} with 200 rows (samples) and 2 columns (sample name & binary tumor size indicator)
}
\usage{
pam50_testdata_samplesheet
}
\description{
Example samplesheet for input samples. Can be used to indicate tumor size value for clinical risk model (ROR-C).
}
\keyword{datasets}
