% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_centroid_heatmap}
\alias{plot_centroid_heatmap}
\title{Plot a heatmap of per sample centroid correlations}
\usage{
plot_centroid_heatmap(input_df, output_png = NULL)
}
\arguments{
\item{input_df}{\link{data.frame} or \link{matrix}.}

\item{output_png}{\link{character}. Optional ouput path to save plot.}
}
\value{
A \pkg{ComplexHeatmap} object
}
\description{
\code{plot_centroid_heatmap} creates a heatmap of centroid correlation for each sample.
}
\details{
Input is the data.frame output of \code{\link[=pam50]{pam50()}}.
}
