% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_pam50_heatmap}
\alias{plot_pam50_heatmap}
\title{Plot a heatmap of PAM50 gene expression across samples}
\usage{
plot_pam50_heatmap(input_df, input_matx, output_png = NULL)
}
\arguments{
\item{input_df}{\link{data.frame}. See \code{\link[=pam50]{pam50()}}}

\item{input_matx}{numeric \link{matrix}. See \code{\link[=preprocess_input]{preprocess_input()}} output.}

\item{output_png}{\link{character}. Optional ouput path to save plot.}
}
\value{
A \pkg{ComplexHeatmap} object
}
\description{
\code{plot_pam50_heatmap} creates a heatmap of centered PAM50 gene expression with subtype and ROR group assignment annotations.
}
\details{
Input is the data.frame output of \code{\link[=pam50]{pam50()}} and processed gene expression from \code{\link[=preprocess_input]{preprocess_input()}}. \pkg{ComplexHeatmap} required.
}
