% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plot_pam50_pca}
\alias{plot_pam50_pca}
\title{PCA plot of PAM50 gene expression across samples.}
\usage{
plot_pam50_pca(input_df, input_matx, output_png = NULL)
}
\arguments{
\item{input_df}{\link{data.frame}. See \code{\link[=pam50]{pam50()}}}

\item{input_matx}{\link{matrix}. See \code{\link[=preprocess_input]{preprocess_input()}} output.}

\item{output_png}{\link{character}. Optional ouput path to save plot}
}
\value{
A \pkg{ggplot2} object
}
\description{
\code{plot_pam50_pca} creates a PCA plot annotated with assignmed PAM50 subtypes. \pkg{ggplot2} required.
}
