% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotPAM50}
\alias{plotPAM50}
\title{Visualize PAM50 Prediction}
\usage{
plotPAM50(es, PAM50.pred, add_labels = NULL, cluster_rows = F,
  cluster_cols = F, scale = "row", ...)
}
\arguments{
\item{es}{dataframe, rows are genes columns are samples. Row names has to be EntrezID.}

\item{PAM50.pred}{array of each sample's PAM50 label
(you can get the label using PAM50 function)}

\item{add_labels}{Additional argument for adding extra labels (dataframe) to show in
the heatmap. Labels have to be the same sample order as PAM50.pred}

\item{cluster_rows}{bool (defaul = False), indicate whether to reorder the genes.
Default uses Parker et al. paper's gene order.}

\item{cluster_cols}{bool (defaul = False), indicate whether to reorder the samples.}

\item{scale}{string (default = "row"). Options are "none", "row", or "column".
This is to rescale data (calculate zscore) based on row, column or no rescaling.}

\item{...}{other parameters passed onto pheatmap, for instance, main = "Treatment Arms".}
}
\description{
This is to generate a heatmap to check the result of PAM50 prediction.
}
