% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotValidationGenes}
\alias{plotValidationGenes}
\title{Plot Validation Genes}
\usage{
plotValidationGenes(es, PAM50.pred, genes = c("ESR1", "PGR", "MKI67",
  "ERBB2"))
}
\arguments{
\item{es}{dataframe, rows are genes columns are samples. Row names has to be EntrezID.}

\item{PAM50.pred}{array of each sample's PAM50 label}

\item{genes}{(optional) vector of validation gene names}
}
\description{
This plots distribution of validation genes (4 by default) for validating
the prediction result.
}
\details{
Default 4 genes:

"ERBB2": Her2 gene; Her2 group should have the highest ERBB2 value.

"ESR1": ER gene; LumA and LumB groups should have the highest ESR1 value.

"MKI67": tumor growth factor; LumA should have the lowest MKI67 value because
LumA patients usually have the best prognosis.

"PGR": PR gene; LumA should have the highest PGR value.
}
