% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{prepare_df}
\alias{prepare_df}
\title{Check input data frame for main gPAM50 function}
\usage{
prepare_df(df)
}
\arguments{
\item{df}{input data frame that is used for prediction}
}
\description{
Check input data frame for main gPAM50 function
}
\author{
Paul Paczuski
}
