% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{preprocess_input}
\alias{preprocess_input}
\title{Preprocess input data matrix}
\usage{
preprocess_input(input_matx, centers = "self", impute = FALSE)
}
\arguments{
\item{input_matx}{\code{matrix}; output of \code{\link[=read_input]{read_input()}}.}

\item{centers}{"self" | \link{matrix} | \link{NULL}; If "self" (default), median center the input directly. If NULL, don't center. Otherwise use custom medians (50 gene x 1 numeric matrix with rownames) for input centering.}

\item{impute}{\code{logical}; impute missing values. \code{\link[impute:impute.knn]{impute::impute.knn()}} required.}
}
\value{
50 genes x n_samples numeric matrix
}
\description{
\code{preprocess_input} prepares the input dataset for comparison to centroids.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
input <-
  read_input(file.path("tests", "testthat", "data", "pam50_testdata.tsv"))
out_matx <- preprocess_input(input)
\dontshow{\}) # examplesIf}
}
