% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_input}
\alias{read_input}
\title{Read input data matrix}
\usage{
read_input(path, delim = "\\t")
}
\arguments{
\item{path}{\link{character}; filesystem path to input file}

\item{delim}{\link{character}; column delimeter used in input file}
}
\value{
50 genes x n_samples numeric \link{matrix}
}
\description{
\code{read_input} returns a 50 x n_samples matrix for comparison to  PAM50 centroids.
}
\details{
The input file should be a plaintext tabular file, with genes in rows and samples in columns. The first column should correspond to HGNC gene symbols and will be treated as \link{data.frame} rownames. Any genes not present in the PAM50 set will be discarded. See \link{pam50_annot}, which includes Ensembl stable gene IDs and HGNC IDs for each PAM50 HGNC gene symbol. This function will panic if not all PAM50 genes are present.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_input(file.path("raw_data", "pam50_testdata.tsv"))
\dontshow{\}) # examplesIf}
}
