% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_tumor_sizes}
\alias{read_tumor_sizes}
\title{Read logical tumor size vector from a tabular samplesheet}
\usage{
read_tumor_sizes(
  path,
  delim = "\\t",
  samplenames = "sample",
  tumor_size_col = "tumor_size"
)
}
\arguments{
\item{path}{\link{character}; filesystem path to input file}

\item{delim}{\link{character}; column delimeter used in input file}

\item{samplenames}{\link{character}; name of column containing samplenames/id. Needs to be same as those in the input expression matrix columns.}

\item{tumor_size_col}{\link{character}; name of column containing tumor size indictor [0, 1] or [FALSE, TRUE].}
}
\value{
N sample named \link{logical} vector of tumor size indicators
}
\description{
Very basic helper func to get named vector of tumor sizes from a samplesheet file.
}
