% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ror_coeffs}
\alias{ror_coeffs}
\title{Risk of Relapse (ROR) model coefficients}
\format{
An object of class \code{list} of length 4.
}
\usage{
ror_coeffs
}
\description{
A nested \link{list} of linear model coefficients for each possible ROR model: subtype only [ROR-S], subtype + tumor size (clinical) [ROR-C], subtype + proliferative score [ROR-P],
and subtype + tumor size + prolif score [ROR-PC]. Taken from legacy
code, see \link{pam50_testdata}.
}
\keyword{datasets}
