% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ror.R
\name{ror_group}
\alias{ror_group}
\title{Assign Rick of Relapse (ROR) group}
\usage{
ror_group(ror_scores)
}
\arguments{
\item{ror_scores}{named \code{numeric}; single set of scores}
}
\value{
Named numeric list of ROR group assignments
}
\description{
\code{ror_group} is an internal helper function to go from raw ROR scores
to 'low', 'medium', or 'high' risk groups
}
