% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ror_thresholds}
\alias{ror_thresholds}
\title{Risk of Relapse (ROR) low vs high thresholds}
\format{
An object of class \code{list} of length 4.
}
\usage{
ror_thresholds
}
\description{
A nested \link{list} of risk score thresholds to call low vs medium vs high risk. See \link{ror_coeffs}
}
\keyword{datasets}
