% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam50.R
\name{pam50}
\alias{pam50}
\title{Run entire PAM50 analysis}
\usage{
pam50(
  input,
  output_dir = NULL,
  centers = "self",
  impute = FALSE,
  tumor_sizes = NA,
  include_normal = TRUE,
  create_qc_figures = FALSE
)
}
\arguments{
\item{input}{\link{character} |  \link{matrix}; See \code{\link[=read_input]{read_input()}}.}

\item{output_dir}{\link{character} | \link{NULL}; directory path for all output. Must already exist.}

\item{centers}{for PAM50 genes. One of:  "self", \link{NULL}, \link{matrix}, or \link{character} path; See \code{\link[=preprocess_input]{preprocess_input()}}}

\item{impute}{\code{logical}; impute missing values in input. Default false. See \code{\link[=preprocess_input]{preprocess_input()}}.}

\item{tumor_sizes}{\link{logical} | \link{character}. Indicator vector of tumor size >2cm for ROR analsis. If a path, will try to read input as TSV with first column of sample names, and another column called tumor_size'. If \link{NA} (default), will ignore.}

\item{include_normal}{\link{logical}; See \code{\link[=assign_subtypes]{assign_subtypes()}}}

\item{create_qc_figures}{\link{logical}. See \code{\link[=plot_cor_box]{plot_cor_box()}}, \code{\link[=plot_pam50_heatmap]{plot_pam50_heatmap()}}, and \code{\link[=plot_pam50_pca]{plot_pam50_pca()}}. Will save to \code{output_dir}.}
}
\value{
A n_sample x 15 features \link{data.frame}. Optionally writes results to \code{output}. Logs to stdout.
}
\description{
\code{pam50} is a helper function to paramaterize and run the full analysis, including: io, centering/scaling, subtype calling, ROR calculations, and (optional) QC plots.
}
